+++ hypr/utils/json.lua
